Model update

+++ LSTM_Model.py
@@ -8,6 +8,7 @@ from keras.layers import LSTM, Dense, TimeDistributed, Conv1D, MaxPooling1D, Fla
 from constants import CLASS_OUTPUT, SEQUENCE_LENGTH, MODEL_INPUT
 from keras.optimizers import Adam
 from keras.losses import categorical_crossentropy as sparse
+import tensorflow as tf
 
 
 # Function create an tensor flow model structure which can be called and trained on
@@ -19,16 +20,16 @@ def create_model(input_shape):
     # layers in the model
     #####################################################################################
 
-    model.add(Reshape((SEQUENCE_LENGTH, MODEL_INPUT, 1), input_shape=input_shape))
-    model.add(TimeDistributed(Conv1D(filters=64, kernel_size=3, activation='tanh')))
-    model.add(TimeDistributed(MaxPooling1D(pool_size=2)))
-    model.add(TimeDistributed(Conv1D(filters=32, kernel_size=3, activation='tanh')))
-    model.add(TimeDistributed(MaxPooling1D(pool_size=2)))
-    model.add(TimeDistributed(Conv1D(filters=16, kernel_size=3, activation='tanh')))
-    model.add(TimeDistributed(MaxPooling1D(pool_size=2)))
-    model.add(TimeDistributed(Flatten()))
-    model.add(LSTM(units=16, return_sequences=True))
-    model.add(Dense(units=CLASS_OUTPUT, activation='softmax'))
+    model.add(Reshape((SEQUENCE_LENGTH, MODEL_INPUT, 1), input_shape=input_shape, dtype=tf.float32))
+    model.add(TimeDistributed(Conv1D(filters=64, kernel_size=3, activation='tanh', dtype=tf.float32)))
+    model.add(TimeDistributed(MaxPooling1D(pool_size=2, dtype=tf.float32)))
+    model.add(TimeDistributed(Conv1D(filters=32, kernel_size=3, activation='tanh', dtype=tf.float32)))
+    model.add(TimeDistributed(MaxPooling1D(pool_size=2, dtype=tf.float32)))
+    model.add(TimeDistributed(Conv1D(filters=16, kernel_size=3, activation='tanh', dtype=tf.float32)))
+    model.add(TimeDistributed(MaxPooling1D(pool_size=2, dtype=tf.float32)))
+    model.add(TimeDistributed(Flatten(dtype=tf.float32)))
+    model.add(LSTM(units=16, return_sequences=True, dtype=tf.float32))
+    model.add(Dense(units=CLASS_OUTPUT, activation='softmax', dtype=tf.float32))
 
     ######################################################################################
